+++ src/main/kotlin/org/rncp/ad/infra/api/AdDto.kt
@@ -9,9 +9,9 @@ data class AdDto(
         val userId: String,
         val name: String,
         val description: String,
-        val hourPrice: Float,
-        val latitude: Float,
-        val longitude: Float,
+        val hourPrice: Double,
+        val latitude: Double,
+        val longitude: Double,
         val state: Boolean,
         val link: String
 ) {
